Explicit error when Packit refuses to create another dist-git MR
PackitAPI.sync_release() no longer returns None (but the MR itself)
in case of already existing MR.
https://github.com/packit/packit/pull/1623

If the returned MR is already in the DB it probably means that
Packit didn't create a new MR and returned an already existing MR.
This should not happen and I actually don't know how to reproduce it.
(https://github.com/packit/hardly/issues/70#issuecomment-1252384033)
But in case it happens again throw an error message.